uploader snmp md file
