+++ README.md
@@ -1,0 +1,32 @@
+# QuecPython SNMP
+
+[中文](./README_ZH.md) | English
+
+## Overview
+
+The Simple Network Management Protocol (SNMP) is a communication protocol used for managing and monitoring network devices in IP networks, applicable to routers, switches, servers, and other devices. Its design is lightweight and highly scalable, supporting various operating systems and network equipment. SNMP uses community strings as a simple means of access control, allowing network administrators to collect and manage device information.
+
+The main SNMP operations include GET, used to request device information; SET, used to modify device configurations; and TRAP, used for devices to report anomalies to the management station. As the protocol has evolved, SNMP has enhanced its security features. The latest version, SNMPv3, supports message integrity, authentication, and encryption to improve communication security.
+
+## Usage
+
+- [API_Reference](./docs/en/API_Reference.md)
+- [Sample Code](./code/snmp_api.py)
+
+## Contribution
+
+We welcome contributions to improve this project! Please follow these steps to contribute:
+
+1. Fork this repository.
+2. Create a new branch (`git checkout -b feature/your-feature`).
+3. Commit your changes (`git commit -m 'Add your feature'`).
+4. Push to the branch (`git push origin feature/your-feature`).
+5. Open a Pull Request.
+
+## License
+
+This project is licensed under the Apache License. For more details, please see the [LICENSE](./LICENSE) file.
+
+## Support
+
+If you have any questions or need support, please refer to the [QuecPython Documentation](https://python.quectel.com/doc) or open an issue in this repository.
