+++ README_ZH.md
@@ -1,0 +1,32 @@
+# QuecPython SNMP协议
+
+中文| [English](./README.md) 
+
+## 概述
+
+简单网络管理协议（SNMP）是一种在IP网络中用于管理和监控网络设备的通信协议，适用于路由器、交换机、服务器等设备。它的设计轻量而且扩展性强，能够支持各种操作系统和网络设备。SNMP使用社区字符串作为一种简单的访问控制手段，允许网络管理员收集和管理设备信息。
+
+SNMP主要操作包括GET，用于请求设备信息；SET，用于修改设备配置；以及TRAP，用于设备向管理站报告异常。随着版本的发展，SNMP增加了安全功能，最新的SNMPv3版本支持消息完整性、认证和加密，以提高通信安全。
+
+## 用法
+
+- [API说明文档](./docs/zh/API说明文档.md)
+- [示例代码](./code/snmp_api.py)
+
+## 贡献
+
+我们欢迎对本项目的改进做出贡献！请按照以下步骤进行贡献：
+
+1. Fork 此仓库。
+2. 创建一个新分支（`git checkout -b feature/your-feature`）。
+3. 提交您的更改（`git commit -m 'Add your feature'`）。
+4. 推送到分支（`git push origin feature/your-feature`）。
+5. 打开一个 Pull Request。
+
+## 许可证
+
+本项目使用 Apache 许可证。详细信息请参阅 [LICENSE](./LICENSE) 文件。
+
+## 支持
+
+如果您有任何问题或需要支持，请参阅 [QuecPython 文档](https://python.quectel.com/doc) 或在本仓库中打开一个 issue。
